refactor: Ecosystem actors index paths
