+++ src/views/Actors/components/ActorItem/ActorItem.stories.tsx
@@ -11,7 +11,7 @@ import type { SocialMediaChannels } from '@ses/core/models/interfaces/socialMedi
 import type { Meta } from '@storybook/react';
 
 const meta: Meta<typeof ActorItem> = {
-  title: 'Fusion/Actor/ActorItem',
+  title: 'Fusion/Views/Ecosystem Actors Index/Actor Item',
   component: ActorItem,
   parameters: {
     chromatic: {
